refactor(centre-aide): supprimer sections inutilisees et extraire SectionWidgetsChantiersSignales (PIL-1370)
