+++ src/client/components/PageAccueil/PageAccueil.tsx
@@ -6,11 +6,8 @@ import {
 } from "@/components/PageAccueil/PageAccueilContext";
 import { BasePageAccueilLayout } from "@/components/PageAccueil/BasePageAccueilLayout";
 import { SectionAvancementMoyen } from "@/components/PageAccueil/sections/SectionAvancementMoyen";
-import { SectionRepartitionTerritoriale } from "@/components/PageAccueil/sections/SectionRepartitionTerritoriale";
-import { SectionRepartitionMeteos } from "@/components/PageAccueil/sections/SectionRepartitionMeteos";
-import { SectionCartographie } from "@/components/PageAccueil/sections/SectionCartographie";
 import { SectionComparaisonTerritoires } from "@/components/PageAccueil/sections/SectionComparaisonTerritoires";
-import { SectionChantiersSignales } from "@/components/PageAccueil/sections/SectionChantiersSignales";
+import { SectionWidgetsChantiersSignales } from "@/components/PageAccueil/sections/SectionWidgetsChantiersSignales";
 import { SectionTableauChantiers } from "@/components/PageAccueil/sections/SectionTableauChantiers";
 import type { getServerSideProps } from "@/pages/accueil/chantier/[territoireCode]/index";
 
@@ -49,35 +46,10 @@ export const PageAccueil: FunctionComponent<PageAccueilProps> = (props) => {
         nombreTotalChantiersAvecAlertes={props.nombreTotalChantiersAvecAlertes}
         territoireCode={props.territoireCode}
       >
-        <main>
-          <div className="py-4 px-0 md:px-4">
-            <div className="fr-grid-row">
-              <div className="fr-col-12 fr-col-lg-7 fr-col-xl-6 flex flex-col">
-                <section className="flex flex-1">
-                  <div className="fr-container fr-p-0 flex flex-1">
-                    <div className="fr-grid-row fr-grid-row--gutters fr-mb-0 fr-mt-0 w-full mr-0 md:mr-4">
-                      <div className="fr-col-12 fr-col-xl-6 flex flex-col items-center pr-0 pt-0">
-                        <SectionAvancementMoyen />
-                      </div>
-                      <div className="fr-col-12 fr-col-xl-6 pr-0 pt-0">
-                        <SectionRepartitionTerritoriale />
-                      </div>
-                    </div>
-                  </div>
-                </section>
-                <div className="mr-0 md:mr-4 xl:mr-0">
-                  <SectionRepartitionMeteos />
-                </div>
-              </div>
-              <div className="fr-col-12 fr-col-lg-5 fr-col-xl-6 xl:pl-2">
-                <SectionCartographie />
-              </div>
-            </div>
-            <SectionComparaisonTerritoires />
-            <SectionChantiersSignales />
-            <SectionTableauChantiers />
-          </div>
-        </main>
+        <SectionAvancementMoyen />
+        <SectionComparaisonTerritoires />
+        <SectionWidgetsChantiersSignales />
+        <SectionTableauChantiers />
       </BasePageAccueilLayout>
     </PageAccueilProvider>
   );
